pyproject.toml
[project]
name = "uni-push"
version = "0.1.0"
description = "Unified push gateway for enterprise bots and webhooks."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite",
    "alembic",
    "arq",
    "asyncmy",
    "asyncpg",
    "email-validator",
    "fastapi",
    "httpx",
    "psycopg[binary]",
    "pymysql",
    "pwdlib[argon2]",
    "pydantic-settings",
    "pyjwt",
    "python-multipart",
    "redis",
    "sqlalchemy",
    "uvicorn[standard]",
]

[project.scripts]
unipush-api = "app.main:run"
seed-admin = "app.scripts.seed_admin:main"
reset-password = "app.scripts.reset_password:main"

[dependency-groups]
dev = [
    "fakeredis[lua]>=2.35.1",
    "mypy>=1.20.2",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "ruff>=0.15.12",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["app/tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
